Refactor shiftSelectors and globalSelectors in SyntheticsStateContext
