+++ src/context/SyntheticsStateContext/selectors/globalSelectors.ts
@@ -13,6 +13,7 @@ export const selectGmMarkets = (s: SyntheticsState) => s.globals.markets.markets
 export const selectUiFeeFactor = (s: SyntheticsState) => s.globals.uiFeeFactor;
 export const selectUserReferralInfo = (s: SyntheticsState) => s.globals.userReferralInfo;
 export const selectChainId = (s: SyntheticsState) => s.globals.chainId;
+export const selectDepositMarketTokensData = (s: SyntheticsState) => s.globals.depositMarketTokensData;
 
 export const selectMinCollateralUsd = (s: SyntheticsState) => s.globals.positionsConstants.minCollateralUsd;
 export const selectMinPositionSizeUsd = (s: SyntheticsState) => s.globals.positionsConstants.minPositionSizeUsd;
